Traipse Alpha 'OpenRPG' {100219-00}
Traipse is a distribution of OpenRPG that is designed to be easy to setup and go. Traipse also makes it easy for
developers to work on code without fear of sacrifice. 'Ornery-Orc' continues the trend of 'Grumpy' and adds fixes to
the code. 'Ornery-Orc's main goal is to offer more advanced features and enhance the productivity of the user.

Update Summary (Patch-2)
New Features:
New Namespace method with two new syntaxes

Fixes:
Fix to Server GUI startup errors
Fix to Server GUI Rooms tab updating
Fix to Chat and Settings if non existant die roller is picked
Fix to Dieroller and .open() used with .vs(). Successes are correctly calculated
Fix to Alias Lib's Export to Tree, Open, Save features
Fix to alias node, now works properly
Fix to Splitter node, minor GUI cleanup

+++ orpg/chat/chat_msg.py
@@ -21,13 +21,13 @@
 # Author: Ted Berg
 # Maintainer:
 # Version:
-#   $Id: chat_msg.py,v 1.15 2006/11/04 21:24:19 digitalxero Exp $
+#   $Id: chat_msg.py,v Traipse 'Ornery-Orc' prof.ebral Exp $
 #
 # Description: Contains class definitions for manipulating <chat/> messages
 #
 #
 
-__version__ = "$Id: chat_msg.py,v 1.15 2006/11/04 21:24:19 digitalxero Exp $"
+__version__ = "$Id: chat_msg.py,v Traipse 'Ornery-Orc' prof.ebral Exp $"
 
 from orpg.orpgCore import *
 from chat_version import CHAT_VERSION
@@ -48,8 +48,7 @@ class chat_msg:
         self.takexml(xml_text)
 
     def __del__(self):
-        if self.chat_dom:
-            self.chat_dom.unlink()
+        if self.chat_dom: self.chat_dom.unlink()
 
     def toxml(self):
         return tostring(self.chat_dom)
